Baby steps towards a 3D sim loop.
