+++ agents_playground/gpu/renderers/landscape_renderer.py
@@ -1,40 +1,22 @@
 import wgpu
+from wgpu import GPURenderPassEncoder, GPURenderPipeline
 import wgpu.backends.wgpu_native
-from agents_playground.cameras.camera import Camera
-from agents_playground.gpu.per_frame_data import PerFrameData
-from agents_playground.gpu.pipelines.pipeline_configuration import PipelineConfiguration
 
-from agents_playground.gpu.renderer_builders.renderer_builder import RendererBuilder
-from agents_playground.gpu.renderer_builders.landscape_renderer_builder import (
-    LandscapeRendererBuilder,
-)
 from agents_playground.gpu.renderers.gpu_renderer import GPURenderer
-from agents_playground.scene import Scene
-from agents_playground.simulation.context import SimulationContextBuilder
-from agents_playground.spatial.matrix.matrix import Matrix
 from agents_playground.spatial.mesh import MeshBuffer, MeshData
 
-
 class LandscapeRenderer(GPURenderer):
-    def __init__(self, builder: RendererBuilder | None = None) -> None:
+    def __init__(self, render_pipeline: GPURenderPipeline) -> None:
         super().__init__()
-        self._render_pipeline: wgpu.GPURenderPipeline
-        self.builder = LandscapeRendererBuilder() if builder is None else builder
+        self._render_pipeline = render_pipeline
 
     @property
-    def render_pipeline(self) -> wgpu.GPURenderPipeline:
+    def render_pipeline(self) -> GPURenderPipeline:
         return self._render_pipeline
-
-    def prepare(
-        self,
-        sim_context_builder: SimulationContextBuilder
-    ) -> None:
-        pc = PipelineConfiguration()
-        self._render_pipeline = self.builder.build(sim_context_builder, pc)
         
     def render(
         self,
-        render_pass: wgpu.GPURenderPassEncoder,
+        render_pass: GPURenderPassEncoder,
         mesh_data: MeshData,
     ) -> None:
         vertex_buffer: MeshBuffer = mesh_data.vertex_buffer.unwrap()
